refactor(sozia): nest server packages under sozia/server/
Move sozia/{api,fusion,inference,registry}/ and sozia/server.py
into sozia/server/ so the import tree reflects sozia.common vs
sozia.server.* as described in the LLD dependency rules.

sozia/server.py becomes sozia/server/__init__.py — existing
`from sozia.server import ...` call sites are unaffected.
