--- sozia/fusion/__init__.py
@@ -1,17 +1,0 @@
-"""sozia.fusion — fusion layer for the Sozia server."""
-
-from sozia.fusion.degraded_mode_handler import DegradedModeHandler, DegradedStatus
-from sozia.fusion.frame_accumulator import FEATURE_DIM, FrameAccumulator
-from sozia.fusion.orchestrator import FusionOrchestrator
-from sozia.fusion.sign_fusion_policy import SignFusionPolicy
-from sozia.fusion.speech_fusion_policy import SpeechFusionPolicy
-
-__all__ = [
-    "DegradedModeHandler",
-    "DegradedStatus",
-    "FEATURE_DIM",
-    "FrameAccumulator",
-    "FusionOrchestrator",
-    "SignFusionPolicy",
-    "SpeechFusionPolicy",
-]
