--- sozia/inference/sign/__init__.py
@@ -1,6 +1,0 @@
-"""sozia.inference.sign — sign-path inference engines."""
-
-from sozia.inference.sign.gloss_to_text_engine import GlossToTextEngine
-from sozia.inference.sign.tsl_recognition_engine import TslRecognitionEngine
-
-__all__ = ["TslRecognitionEngine", "GlossToTextEngine"]
