+++ sozia/server/fusion/__init__.py
@@ -1,0 +1,17 @@
+"""sozia.fusion — fusion layer for the Sozia server."""
+
+from sozia.server.fusion.degraded_mode_handler import DegradedModeHandler, DegradedStatus
+from sozia.server.fusion.frame_accumulator import FEATURE_DIM, FrameAccumulator
+from sozia.server.fusion.orchestrator import FusionOrchestrator
+from sozia.server.fusion.sign_fusion_policy import SignFusionPolicy
+from sozia.server.fusion.speech_fusion_policy import SpeechFusionPolicy
+
+__all__ = [
+    "DegradedModeHandler",
+    "DegradedStatus",
+    "FEATURE_DIM",
+    "FrameAccumulator",
+    "FusionOrchestrator",
+    "SignFusionPolicy",
+    "SpeechFusionPolicy",
+]
