+++ sozia/server/fusion/degraded_mode_handler.py
